[skip changelog] Restore globbing of nightly build artifact filename
During a refactoring of the "Publish Nightly Build" workflow, globbing of the macOS build artifact filename in the
repackaging step was accidentally lost, causing the glob pattern to be treated instead as a string and the updated
package saved to a file named by that string instead of to the original filename.

This resulted in the workflow failing with errors like:

Artifact path is not valid: /arduino-cli_nightly-*macOS_64bit.tar.gz. Contains the following character: Asterisk *

The previous `basename` command was providing the globbing. After the refactoring, `basename` is no longer needed, so an
alternative way to achieve globbing is needed. It seems the preferred approach is use of an array. Since the globbing
will only expand to a single filename, the array can be referenced as before by the rest of the step, since it will
resolve to the first element in the array (equivalent to `$PACKAGE_FILENAME[0]`).