Load package indexes after installed packages
The installed packages keeps an extract of the original
package_index.json, at the moment of the installation, inside
a file named installed.json. This extract may turn out useful
if the original package_XXX_index.json is lost or removed to keep the
platform functional. On the other hand, if the original
package_XXX_index.json is modified the information kept in the
installed.json may be outdated and should be replaced by the upstream
index: this is the reason why it's loaded after the hardware platforms.