[skip changelog] Enable Codecov comments on PRs from forks
Versions of the `codecov/codecov-action` GitHub Actions action prior to 1.0.6 required the use of a token provided by
Codecov in order to upload data to Codecov. This token was stored in secret in the Arduino CLI repository and used in
the test workflow.

For security reasons, secrets are not accessible when a workflow is triggered by an event generated by a fork of the
repository. This meant that it was impossible to upload coverage data for the test runs triggered by PRs from forks. A
conditional was added to the upload step of the workflow to cause it to only run on `push` event triggers, which
effectively prevented its failure for runs on PRs from forks.

The token requirement was removed in the 1.0.6 release of `codecov/codecov-action`, but the now pointless conditional
was never removed from the workflow. This prevented PRs from forks from receiving the automated code coverage report
comments that would otherwise encourage those contributors to resolve coverage deficiencies and facilitate the review
process.

The harmful conditional is hereby removed from the coverage data upload steps of the workflow and PRs from forks will
now receive coverage report comments, just as PRs from branches do already.