[skip changelog] Remove unused link reference definitions in docs
Parts of the documentation use reference links, where the inline markup uses a label, which matches to a link reference
definition elsewhere in the documentation.

The major disadvantage of this approach is that unused link reference definitions are often left behind after the
referencing text is removed, cluttering up the documentation and increasing the maintenance burden. That was the case
here.

The unused link reference definitions are hereby removed.