[skip changelog] Update documentation links that depend on redirects
Some of the links in the documentation used URLs that were redirecting to the new location of the target page.

Even though these links still took the reader to the intended place eventually, it is safest to not rely on them
continuing to work for the long term, so is best to update to point directly to the target.