[skip changelog] Update broken or outdated links (#1833)
* [skip changelog] Remove unused link reference definitions in docs

Parts of the documentation use reference links, where the inline markup uses a label, which matches to a link reference
definition elsewhere in the documentation.

The major disadvantage of this approach is that unused link reference definitions are often left behind after the
referencing text is removed, cluttering up the documentation and increasing the maintenance burden. That was the case
here.

The unused link reference definitions are hereby removed.

* [skip changelog] Fix broken links in documentation

Some of the documentation links no longer reached their intended target.

These were not caught by the Markdown link check for one of two reasons:

- Redirects to a useless generic page instead of the moved specific target page
- Links in non-Markdown files

* [skip changelog] Update documentation links that depend on redirects

Some of the links in the documentation used URLs that were redirecting to the new location of the target page.

Even though these links still took the reader to the intended place eventually, it is safest to not rely on them
continuing to work for the long term, so is best to update to point directly to the target.