[skip changelog] Add readme for localization data
Arduino CLI has been translated to 12 languages.

The localization process follows the following steps:

1. An English language source string is defined in the Arduino CLI codebase
2. The source string is pushed to Transifex
3. Community translators localize the string
4. The localization data is pulled into the Arduino CLI repository
5. The localization data is incorporated into the Arduino CLI distribution

Experience with maintenance of the Arduino IDE 1.x project indicates that the data files generated at step (4) can
appear to be the appropriate place to make edits for casual contributors not familiar with the project's sophisticated
internationalization infrastructure.

Since those files are generated by automated systems, any edits made there would only be overwritten, so it is important
to clearly communicate the correct way to make enhancements or corrections to these strings. This is accomplished by a
local readme file most likely to be seen by those working in the folder containing these files, which supplements the
existing information about translation in the project's CONTRIBUTING.md.