Correct key name in `arduino-cli compile --dump-profile` output (#1846)
The `--dump-profile` flag causes Arduino CLI to print the board and sketch dependencies of that compilation as a YAML
document. This could be copied into a "sketch project file" for later use as a "build profile" specified via the
`--profile` flag.

A typo in the `profiles` key name caused the generated "build profile" entry to not be recognized when added to a
"sketch project file".

The key name is hereby corrected.