[skip-changelog] Set test environment directory as CLI WorkingDir (#1886)
* Set test environment directory as CLI WorkingDir

By default, the working directory is the one containing the test.go file. This causes problems when executing commands that have to create files specifically in the working directory, because they either must be deleted manually or the user has to be aware of it and defer a deleting instruction. Furthermore, it messes with tests using relative paths.
Setting the environment directory as the CLI's WorkingDir prevents the above mentioned issues from occurring.

* Fix errors related to the change of the working directory

* Use absolute path to create daemon environment