[skip-changelog] Use a matrix to create parallel builds for each OS (#1883)
* Use a matrix create parallel builds for each os

Using a matrix to run each build task greatly improves performances,
since they can all start concurrently. The finishing time of the job
will be equal to the one of the longer task to build, instead of being
the sum of each individual task's finishing time.

* Calculate checksums during release creation

Checksums of the output files where previously calculated during the initial creation of the artifacts, during the notarization process and, finally, at the release creation's step.
The whole process has been simplified and checksums are now computed only during the creation of the release.

* Set condition to create changelog once

The changelog is the same for each OS. It does not make sense to generate it more than once.

* Disable s3 push for testing

* Upload nightly artifacts for testing

* Fix linux_arm_6 typo

* Stop uploading nigthly artifacts

* Enable s3 pushing

* Upload build artifacts separately

Previously, the different builds were firstly uploaded using a single artifact, which was then downloaded to create different ones and eventually deleted.
Now, since builds are created concurrently, the same matrix can be used to directly upload an artifact for each build. It's necessary to use a second job to calculate the checksum related to
each build and save them all in a single .txt file.