[skip changelog] Use correct markup for PR template headings
This repository uses a pull request template in order to provide a standardized structure for pull request messages.

The template imposes a framework for the organization of the pull request message by providing a section to contain each
of the distinct types of important information to be included in the pull request. As should be obvious, the only
correct formatting for section headings is heading markup, yet somehow list markup was chosen instead. This collides
with the appropriate use of lists within the sections.

The template is hereby changed to use the correct markup for headings.