[skip changelog] Fix pull request template markup (#1901)
* [skip changelog] Use correct markup for PR template headings

This repository uses a pull request template in order to provide a standardized structure for pull request messages.

The template imposes a framework for the organization of the pull request message by providing a section to contain each
of the distinct types of important information to be included in the pull request. As should be obvious, the only
correct formatting for section headings is heading markup, yet somehow list markup was chosen instead. This collides
with the appropriate use of lists within the sections.

The template is hereby changed to use the correct markup for headings.

* [skip changelog] Move contributor guide link to appropriate section of PR template

This repository uses a pull request template in order to provide a standardized structure for pull request messages.

The template includes a link to the project's contributor guide. Previously, this link was at the bottom of the
template, separated from the other content by a horizontal rule. One of the unfortunate ambiguities of the Markdown
syntax is that the horizontal rule syntax is the same as the "setext" H2 heading syntax, with the differentiation
depending on whether the prior line is empty. This results in text intended to be part of the "Other information"
section of the PR message instead being formatted as an H2 heading if the contributor does not add an empty line at the
end of the "Other information" section.

Moving the link to the section of the template that contains a checklist for the PR requirements makes the relevant
information readily accessible to the contributor while also avoiding the creation of a formatting trap.