Move upload of coverage data to dedicated job in test workflow
The code coverage data generated by running the unit tests in the "Test Go" GitHub Actions workflow is uploaded to
Codecov.

This upload is prone to occasional transient failures. With the previous workflow configuration this caused a failure of
the Linux test run job. Contributors would need to review the logs to understand that these failures were not caused by
a test failing. The failure also causes the ongoing jobs that run the tests on other operating systems to be immediately
canceled, meaning their full test results are not available. The entire test suite must be reran just to attempt the
upload again.

Moving the upload to a dedicated workflow job makes it faster and easier for contributors to interpret the cause of a
test failure. It also allows the test suite to complete for all operating systems, making their results immediately
available to the contributor even when the coverage data upload fails. The coverage upload job can be reran, making
recovery from a transient failure more efficient.