Use a single step to upload coverage data in "Test Go" workflow
The code coverage data generated by running the unit tests in the "Test Go" GitHub Actions workflow is uploaded to
Codecov. Arduino CLI's unit tests are split into two collections: those covering the modernized part of the codebase,
and those covering the "legacy" code inherited from arduino-builder. A separate code coverage data file is produced for
each of these collections.

The `codecov/codecov-action` GitHub Actions action is used to perform this upload. Previously, a separate step was used
for the upload of each of the code coverage files. The action provides the capability to specify multiple files for
upload in a comma-separated list. This approach will make the workflow more efficient and reliable.