[skip-changelog] Add ARMv6 download links to install docs
As a newcomer, this made me assume ARMv6 is not supported so I searched the issues, found https://github.com/arduino/arduino-cli/issues/340, which suggested building from source so I did, only to discover an ARMv6 binary in the release files.

As for the table layout, it's not ideal. It would also make sense to have this in another column.