Run integration tests concurrently
This enhancement reduces the amount of time needed to execute integration tests, since each test package no longer has to wait for the previous ones to be completed. In order to do this, a regex is used
to match each test file. It is also specified the path to the directory that contains the file.