[skip changelog] Use correct plurality for word in error message (#1954)
The `lib search` command accepts search keywords as an argument. All libraries with metadata within the search scope
matching those keywords will be returned.

Previously, the error message used the singular form "library". Since the command is as likely to be used to search for
libraries, using the plural form "libraries" is more appropriate.

Users opened a task inside Transifex requesting this change.

Co-authored-by: per1234 <accounts@perglass.com>