[skip changelog] Document filter setting capability of Library/Boards Manager links (#1958)
Arduino IDE allows opening searches in Library Manager and Boards Manager by clicking a link with a special format in a
comment of the sketch.

During the initial implementation in Arduino IDE 1.x, support for setting the "Type" filter via the URI in addition to
specifying search keywords was attempted and claimed to work, but never did work. For this reason, it was not previously
mentioned in the sketch specification.

During the reimplementation of the Library/Boards Manager link feature for Arduino IDE 2.x, the "Type" filter setting
capability was finally truly implemented, and expanded to also support setting the "Topic" filter in Library Manager via
the same mechanism.

Now that support for the filter setting capability has been released in Arduino IDE 2.0.1, it should be documented in
the sketch specification.