[skip changelog] Add missing checkout steps to build publishing workflow jobs
GitHub Actions workflows are used to automatically generate nightly builds and production releases of the project.

A workflow job is dedicated to the publishing of the build. Previously, this job only needed to download the workflow
artifact containing the build filess created by the previous job, generate a checksum file, and upload the builds to
Arduino's server. None of these operations had any need for the contents of the repository, so a checkout step was not
added to the job.

An additional operation was recently added to the publishing job: the generation of archives containing the protocol
buffer files to publish along with the builds. This new operation does require the contents of the repository. A
checkout step was not added to the job at that time, which caused the protocol buffer generation step and the publishing
job to fail:

task: No Taskfile found in "/home/runner/work/arduino-cli/arduino-cli". Use "task --init" to create a new one

The problem is resolved by adding a step that uses the `actions/checkout` GitHub Actions action to checkout the Arduino
CLI repository into the runner machine to make it available for use in that job.