[skip changelog] Don't error protobuf archive task if distribution folder already exists
The `protoc:collect` task generates a ZIP archive of the project's protocol buffer files for inclusion with the
published builds.

In the case of the nightly and release build workflows, the distribution folder the archive is generated into was
already created by the build operations in the preceeding workflow job. Previously, the task (and thus entire build
workflow) would fail in this case when attempting to create the folder that already exists:

mkdir: cannot create directory ‘../dist’: File exists

In addition to the obvious purpose, the `--parents` flag of the `mkdir` command also has the following additional
effect:

> no error if existing

So adding this flag to the command resolves the spurious failure. It will create the folder when necessary, but simply
carry on with the rest of the task when it does exist.