[skip changelog] Set correct version for nightly build protobuf archive filename
Different version identifiers are used for the build archive filenames depending on the type of build:

Local: git-snapshot
Tester: test-<reference>-git-snapshot
Nightly: nightly-<date>
Production: <tag>

The use of a nightly build version format is specified by setting the `NIGHTLY` environment variable to `true` in the
GitHub Actions workflow.

An additional step was recently added to the build workflows: the generation of archives of protocol buffer files to
publish along with the builds. The setting of the `NIGHTLY` environment variable in that step was neglected in the
nightly build workflow, which caused the protocol buffer archive file to be named with the local build version
identifier instead of the nightly build version identifier. This also resulted in the protocol buffer archive file not
being added to the checksums file.