[skip changelog] Remove comment that breaks gRPC documentation structure (#2080)
The "gPRC Reference" section of the documentation website is automatically generated from the code and comments in the
project's Protocol Buffers files. The documentation is automatically structured according to the structure of the
Protocol Buffers code using heading levels, and the documentation website uses these heading levels to generate a table
of contents to allow easy navigation of the documentation and an overview of its structure.

The generation happens in two steps:

Protocol Buffers -> Markdown -> HTML

A comment in a Protocol Buffers file included a decorative underline formed from a series of `-` characters. This
happens to be markup for an H2 heading in the Markup language. This caused an inadvertent creation of an inappropriate
heading named "BOOTSTRAP COMMANDS", which resulted in the documentation having the following incorrect structure:

- Protocol Documentation
  - [...]
  - cc/arduino/cli/commands/v1/board.proto
    - [...]
  - cc/arduino/cli/commands/v1/commands.proto
    - [...]
    - ArduinoCoreService
  - BOOTSTRAP COMMANDS
    - <List of cc.arduino.cli.commands.v1.ArduinoCoreService methods>
  - cc/arduino/cli/commands/v1/common.proto
    - [...]
  - [...]

Instead of the correct structure:

- Protocol Documentation
  - [...]
  - cc/arduino/cli/commands/v1/board.proto
    - [...]
  - cc/arduino/cli/commands/v1/commands.proto
    - [...]
    - ArduinoCoreService
      - <List of cc.arduino.cli.commands.v1.ArduinoCoreService methods>
  - cc/arduino/cli/commands/v1/common.proto
    - [...]
  - [...]

This could be corrected by removing only the pointless decorative comment, leaving the "BOOTSTRAP COMMANDS" comment, but
I don't see any value in that comment and it introduced unpleasant caps lock prose content into the documentation so I
removed it as well.