Explain usage of `monitor --config` in command help
The `--config` flag of the `arduino-cli monitor` command is used to configure the communication port used by the
monitor.

Previously the command line help did not provide any guidance for the usage of this flag, which might lead the users to
wonder:

- How can the configuration options available for use via the flag be determined?
- What is the format for the configuration option argument?

The information is provided in the FAQ page of the documentation, but that is not as convenient a source of information
as the command line help and the user may not even be aware of its existence.

The command help is hereby adjusted to provide the user with this information:

- Create a conceptual linkage between the `--config` and `--describe` flags by using the "communication port settings"
  terminology in the references of both.
- Document the argument format.

In addition to the comma-separated list format I documented here, an alternative usage of passing multiple `--config`
flags is also supported. I chose the former because I felt that the descriptions of the latter in either command line
notation (e.g., `[--config <ID>=<value>]...`) or in prose (e.g., "The format is <ID>=<value>. Can be used multiple times
for multiple settings.") were less clear.