cr: remove `latest_compatible_version` from gRPC api.
We leave only the `latest_compatible` field from the API. This field
will have the same meaning the the former `latest_compatible_version`.
Meaning that if the `latest_compatible` contains a version it is a
release that can be installed. However it might happen that we have
newer releases that cannot be installed. If the client is interested to
show newer-non-installable-releases, the `releases` field can be used to
infer that.