Add [skip-changelog] prefix to dependabot commit
We use the `[skip-changelog]` prefix to avoid having such commits in our
changelog history.
When merging dependabot PRs we always manually add the prefix, but it's
common to forget about it at costs of manually removing that entry in
the changelog when publishing a new release.

With this PR dependabot wil add that prefix in every commit, the GH PR
will be open with the same commit name, and when pressing the `merge squash`
we won't need to manually add the prefix.