[skip changelog] Correct format of `upload_port` identification properties in snippet (#2655)
The platform developer can associate port properties with a board definition in order to cause Arduino CLI to identify
ports having those properties as that board.

The list format is supported for these platform properties in order to allow multiple alternative port property sets to
be associated. The platform property must have the format `upload_port.n.<property ID>` (where "n" is the array index of
the port property set).

Although the platform property format is correctly documented in the Arduino Platform Specification, an incorrect format
`upload_port.<property ID>.n` was previously used in the boards.txt snippet illustrating the use of the platform
properties.