Document breaking change to `config get` command output
Previously, the `config dump` CLI command returned the effective configuration, including both the values explicitly set
via the configuration file and environment variables as well as the values provided by defaults.

It was changed to only return the explicitly set configuration data. The breaking change was not documented at that time.